Adición modelo de respuesta users. Put y delete JWT

+++ app/controllers/user_controller.py
@@ -1,12 +1,24 @@
 from flask import request, jsonify
 from flask_restx import Namespace, Resource, fields
 from app.services.user_service import UserService
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 # Crear un espacio de nombres (namespace) para los usuarios
 user_ns = Namespace('users', description='Operaciones relacionadas con los usuarios')
 
-# Definir el modelo de usuario para la documentación de Swagger
+# Modelo de entrada para usuarios
 user_model = user_ns.model('User', {
+    'email': fields.String(description='Correo electrónico del usuario'),
+    'password': fields.String(description='Contraseña del usuario'),
+    'username': fields.String(description='Nombre de usuario de identificación'),
+    'name': fields.String(description='Nombre del usuario'),
+    'bio': fields.String(description='Biografía del usuario'),
+    'profile_pic': fields.String(description='Foto de perfil del usuario')
+})
+
+# Modelo de salida (respuesta) para usuarios
+user_response_model = user_ns.model('UserResponse', {
+    'id_user': fields.Integer(description='ID del usuario'),
     'email': fields.String(description='Correo electrónico del usuario'),
     'password': fields.String(description='Contraseña del usuario'),
     'username': fields.String(description='Nombre de usuario de identificación'),
@@ -21,6 +33,7 @@ user_model = user_ns.model('User', {
 class UserResource(Resource):
     @user_ns.doc('create_user')
     @user_ns.expect(user_model, validate=True)  # Decorador para esperar el modelo en la petición
+    @user_ns.marshal_with(user_response_model, code=201)  # Serialización automática del usuario creado
     def post(self):
         """
         Crear un nuevo usuario
@@ -48,10 +61,10 @@ class UserResource(Resource):
             if field not in data: 
                 return jsonify({'error': f'El campo {field} es requerido'}), 400
 
-        user = UserService.create_user(data['email'], data['password'], data['username'],
-                                       data['name'], data.get('bio'), data.get('profile_pic'), data.get('member_since'))
+        user = UserService.create_user(data)
         # Usamos jsonify para asegurarnos de que la respuesta siga el formato JSON válido.
-        return jsonify({'message': 'User created successfully', 'user': user.username})
+        #return jsonify({'message': 'User created successfully', 'user': user.username})
+        return user
 
     @user_ns.doc('get_users')
     def get(self):
@@ -71,6 +84,7 @@ class UserResource(Resource):
 @user_ns.route('/<username>')
 @user_ns.param('username', 'El nombre de usuario')
 class UserDetailResource(Resource):
+    @jwt_required()
     @user_ns.doc('delete_user')
     def delete(self, username):
         """
@@ -85,12 +99,21 @@ class UserDetailResource(Resource):
         - 200: Usuario eliminado con éxito.
         - 404: Si el usuario no se encuentra.
         """
+        id_user = get_jwt_identity() # Obtiene el ID del usuario autenticado
+        user = UserService.get_user_by_username(username)
+
+        # Verifica si el usuario existe y pertenece al usuario autenticado
+        if user.id_user != id_user:
+            return jsonify({'message': "You're not authorized to delete this user"}), 403
+        
         UserService.delete_user(username)  # Llama al servicio para eliminar al usuario
         # Usamos jsonify para enviar un mensaje de éxito en formato JSON.
         return jsonify({'message': 'User deleted successfully'})
 
+    @jwt_required()
     @user_ns.doc('update_user')
     @user_ns.expect(user_model, validate=True)
+    @user_ns.marshal_with(user_response_model, code=201)  # Serialización automática del usuario actualizado
     def put(self, username):
         """
         Actualizar un usuario
@@ -113,7 +136,15 @@ class UserDetailResource(Resource):
         - 200: Usuario actualizado con éxito.
         - 404: Si el usuario no se encuentra.
         """
+        id_user = get_jwt_identity() #Obtiene el ID del usuario autenticado
+        user = UserService.get_user_by_username(username)
+
+        #Verifica si el usuario existe y pertenece al usuario autenticado
+        if user.id_user != id_user:
+            return jsonify({'message': "You're not authorized to update this user"}), 403
+        
         new_data = request.get_json()  # Obtiene los nuevos datos para la actualización
-        UserService.update_user(username, new_data)  # Llama al servicio para actualizar el usuario
+        updated_user = UserService.update_user(username, new_data)  # Llama al servicio para actualizar el usuario
         # Usamos jsonify para enviar un mensaje de éxito en formato JSON.
-        return jsonify({'message': 'User updated successfully'})
+        # return jsonify({'message': 'User updated successfully'})
+        return updated_user, 200
